wake words: the review pass - wrap-proof deadlines, a stall bound, and four smaller escapes
A code review of the runtime loader before this branch settles. Seven
findings, worst first.

Every deadline moves to millis_64(). The loader compared plain millis()
against stored deadlines, and the 32-bit counter wraps at 49.7 days - these
devices stay up for months. The worst case was the handler's wake-test
window: set just before the wrap, it read as open for the next 49 days - a
device that records wake words but answers none of them until reboot. The
milder cases silently stalled boot re-downloads and Home Assistant
reconciliation for the same stretch. The 64-bit clock is thread-safe in the
pinned ESPHome and already in use one file over, so every deadline
(next_retry_ms, tune_deadline_ms_, reconcile_after_ms_, wake_test_until_)
now rides it; the event and detection ages keep their wrap-safe unsigned
subtraction.

A server that accepted the connection and then went quiet parked the
download task forever: read() returning zero on a sized body was waited out
without a bound, the job never finished, and the single job slot - which
every swap and boot re-download shares - was gone until reboot. WL_STALL_MS
(30 seconds without a byte) bounds it in both the model loop and fetch_,
which also stops a stalled manifest read from misreporting as
"not a manifest".

One bad URL locked the tuner out. A failed swap leaves the previous word
listening but the slot in ERROR, and both the tuner row (st === 0) and
apply_tune_ (state != READY) keyed on state - so the working word stayed
untunable until a reboot. The payload gains ld ("a model is loaded and
listening"), the tuner keys on that at both ends, and the same flag stops
the "device retries on its own" note from showing over a failed swap whose
word never stopped working.

The duplicate-word guard re-checks on the main loop. queue_slot checks the
published view from the httpd task, but two requests queued back to back
both pass before either applies - and two slots holding one model breaks
the enabled-iff-in-a-slot invariant every unload and reconcile pass relies
on (emptying one slot would disable the model the other still claims, and
reconciliation would then empty that one too). apply_request_ refuses the
second with ERR_REFUSED, and setup() empties slot 1 if both ever persisted
the same spec.

A queued tune request could be dropped: the tune exchange ran outside
req_lock_, so a request arriving between it and the req_pending_
recomputation was overwritten to "nothing pending" - a lost close left the
model on the probe floor for the whole TTL. The exchange and queue_tune's
store now share the lock.

The pairing write stops trusting a stopwatch. syncAssist fired one blind
select write behind sleep(8000), hoping Home Assistant had finished the
reconnect the nudge forces; a slower reconnect lost the write silently -
the exact bug the function was written to kill, back through a different
door. It now writes, refreshes, verifies against the fresh payload and
retries, bounded at 30 seconds; repeating a landed write is harmless.

WW_ERR gains its 0 line: a write the device refused outright (a URL past
the 224-byte slot spec) or that never got an answer showed ww_failed with a
blank where the reason goes.

Deferred knowingly (design calls with self-healing behavior): a room
false-fire during the tuner's speak phase still counts as an attempt, and
two concurrent tune open/close posts can briefly zero the test window until
the next 20-second keepalive restores it.

+++ esphome/components/mww_runtime_loader/mww_runtime_loader.h
@@ -221,8 +221,9 @@ class MwwRuntimeLoader : public Component {
     uint8_t state{SLOT_READY};
     uint8_t error{ERR_NONE};
     std::string model_id;  // id of the model this slot has loaded/enabled, "" when none
-    // Boot re-download bookkeeping.
-    uint32_t next_retry_ms{0};
+    // Boot re-download bookkeeping. millis_64() so a deadline can never sit on the wrong side of
+    // the 49.7-day millis() wrap - these devices stay up for months.
+    uint64_t next_retry_ms{0};
     uint8_t retries{0};
   };
 
@@ -292,7 +293,8 @@ class MwwRuntimeLoader : public Component {
   // task produced the line (the two lines it parses happen to come from the main loop, but the
   // guard must be safe for every line that merely shares the tag).
   std::atomic<int8_t> tune_slot_{-1};
-  uint32_t tune_deadline_ms_{0};
+  // millis_64(): wrap-proof, like every deadline here.
+  uint64_t tune_deadline_ms_{0};
   // The tuned word and the ring, shared between the log callback's writer and the endpoint's
   // reader. tune_word_ is only written at session open, under the same lock the callback compares
   // it under.
@@ -308,7 +310,8 @@ class MwwRuntimeLoader : public Component {
   Job job_;
   // Set when reconcile_ha_ should hold off for a beat: right after boot (models are still
   // restoring) and right after our own writes (enable() lands on the next inference pass).
-  uint32_t reconcile_after_ms_{0};
+  // millis_64(): wrap-proof, like every deadline here.
+  uint64_t reconcile_after_ms_{0};
   // First-loop re-assertion of built-in cutoff overrides - see loop() for why not setup().
   bool overrides_applied_{false};
 };
